requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==3.2.9
idna==3.4
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
stripe==5.0.0
typing_extensions==4.4.0
urllib3==1.26.12